Update README.md
You recently got hired as a developer in a micro-finance startup with a mission to empower and provide opportunities for low income individuals. The core team currently uses Subversion (SVN) for managing code. They want to slowly move their code to Git. As the first step in this journey, you are asked to host their sample code to calculate simple interest on GitHub in a new repository. You will not only host the script, but also follow best practices introduced in this course and create supporting documents for the open-source project including a code of conduct, and contribution guidelines. Additionally, the repository should be available to the community under the Apache License 2.0.